Add configuration description rules
Merge group rules into general rules files
