+++ BlockServer/block_server.py
@@ -50,7 +50,7 @@ from BlockServer.core.runcontrol_manager import RunControlManager
 from BlockServer.epics.archiver_manager import ArchiverManager
 from BlockServer.core.block_cache_manager import BlockCacheManager
 from BlockServer.site_specific.default.block_rules import BlockRules
-from BlockServer.site_specific.default.group_rules import GroupRules
+from BlockServer.site_specific.default.general_rules import GroupRules, ConfigurationDescriptionRules
 
 
 # For documentation on these commands see the accompanying block_server.rst file
@@ -236,6 +236,7 @@ class BlockServer(Driver):
         self.bumpstrip = "No"
         self.block_rules = BlockRules(ca_server)
         self.group_rules = GroupRules(ca_server)
+        self.config_desc = ConfigurationDescriptionRules(ca_server)
 
         # Connect to version control
         try:
