+++ BlockServer/site_specific/default/test_modules/group_rules_tests.py
@@ -13,7 +13,7 @@
 #along with this program; if not, you can obtain a copy from
 #https://www.eclipse.org/org/documents/epl-v10.php or 
 #http://opensource.org/licenses/eclipse-1.0.php
-from BlockServer.site_specific.default.group_rules import GroupRules
+from BlockServer.site_specific.default.general_rules import GroupRules
 from server_common.mocks.mock_ca_server import MockCAServer
 import unittest
 import json
